Zero out all unused areas of the update vectors
In several places the accelerations portion of the update vectors was
not being zeroed out before the sum of the vector was taken.
For example, if any of the accelerations were marked as "true" on an
odom topic, the poseUpdateSum would be greater than 0 leading to taking
some incorrect code paths (such as calling `poseCallback()`).